requirements.txt
tensorflow
numpy
opencv-python
pillow
scikit-learn